requirements.txt
blinker==1.9.0
click==8.2.1
colorama==0.4.6
contourpy==1.3.3
cycler==0.12.1
filelock==3.13.1
Flask==3.1.2
fonttools==4.59.1
fsspec==2024.6.1
itsdangerous==2.2.0
Jinja2==3.1.4
joblib==1.5.1
kiwisolver==1.4.9
MarkupSafe==2.1.5
matplotlib==3.10.5
mpmath==1.3.0
networkx==3.3
numpy==2.1.2
opencv-contrib-python==4.12.0.88
opencv-python==4.12.0.88
packaging==25.0
pandas==2.3.2
pillow==11.0.0
pyparsing==3.2.3
python-dateutil==2.9.0.post0
pytz==2025.2
scikit-learn==1.7.1
scipy==1.16.1
seaborn==0.13.2
six==1.17.0
sympy==1.13.1
threadpoolctl==3.6.0
torch==2.5.1+cu121
torchaudio==2.5.1+cu121
torchvision==0.20.1+cu121
typing_extensions==4.12.2
tzdata==2025.2
Werkzeug==3.1.3
opencv-python-headless==4.8.0.74 
fastapi==0.103.1 
uvicorn==0.23.2